[rename] sync -> fetch
The sync name is misleading. This task only fetches the information
from the website and makes us aware of any new rules available.
